Bind debug for more test types
