+++ build_test.mk
@@ -24,7 +24,6 @@ GTEST_INTERNAL_INC :=\
 
 $(GTEST_OUTPUT)_SRC :=\
 	googletest/src/gtest-all.cc\
-	googletest/src/gtest_main.cc\
 	googlemock/src/gmock-all.cc
 
 $(GTEST_OUTPUT)_DEFS :=
@@ -57,7 +56,13 @@ include $(QUANTUM_PATH)/sequencer/tests/rules.mk
 include $(QUANTUM_PATH)/serial_link/tests/rules.mk
 ifneq ($(filter $(FULL_TESTS),$(TEST)),)
 include build_full_test.mk
+else
+$(TEST)_SRC += $(LIB_PATH)/printf/printf.c
+$(TEST)_SRC += $(COMMON_DIR)/printf.c
+VPATH += $(LIB_PATH)/printf
 endif
+
+$(TEST)_SRC += tests/test_common/main.c
 
 $(TEST_OBJ)/$(TEST)_SRC := $($(TEST)_SRC)
 $(TEST_OBJ)/$(TEST)_INC := $($(TEST)_INC) $(VPATH) $(GTEST_INC)
